refactor(deployment): move deployment files into subdir
