+++ lib/manager/components/deployment/CurrentDeploymentPanel.js
@@ -9,11 +9,11 @@ import {
   Label as BsLabel
 } from 'react-bootstrap'
 
-import * as deploymentActions from '../actions/deployments'
-import { formatTimestamp } from '../../common/util/date-time'
-import { getActiveInstanceCount, getServerForId } from '../util/deployment'
+import * as deploymentActions from '../../actions/deployments'
+import { formatTimestamp } from '../../../common/util/date-time'
+import { getActiveInstanceCount, getServerForId } from '../../util/deployment'
 import DeploymentPreviewButton from './DeploymentPreviewButton'
-import EC2InstanceCard from '../../common/components/EC2InstanceCard'
+import EC2InstanceCard from '../../../common/components/EC2InstanceCard'
 
 import type {
   Deployment,
@@ -22,7 +22,7 @@ import type {
   OtpServer,
   Project,
   ServerJob
-} from '../../types'
+} from '../../../types'
 
 type Props = {
   deployJobs: Array<ServerJob>,
