+++ lib/manager/components/deployment/DeploymentConfirmModal.js
@@ -4,15 +4,15 @@ import React, {Component, type Node} from 'react'
 import {Alert as BootstrapAlert, Button, Glyphicon, Modal} from 'react-bootstrap'
 import {Map, TileLayer, Rectangle} from 'react-leaflet'
 
-import * as deploymentActions from '../actions/deployments'
-import {getComponentMessages} from '../../common/util/config'
-import {defaultTileLayerProps} from '../../common/util/maps'
-import {getServerForId} from '../util/deployment'
-import {getFeedNames, versionHasExpired} from '../util/version'
+import * as deploymentActions from '../../actions/deployments'
+import {getComponentMessages} from '../../../common/util/config'
+import {defaultTileLayerProps} from '../../../common/util/maps'
+import {getServerForId} from '../../util/deployment'
+import {getFeedNames, versionHasExpired} from '../../util/version'
 import polygon from 'turf-polygon'
 import area from '@turf/area'
 
-import type {Deployment, Project, SummarizedFeedVersion} from '../../types'
+import type {Deployment, Project, SummarizedFeedVersion} from '../../../types'
 
 type Props = {
   deployToTarget: typeof deploymentActions.deployToTarget,
