+++ lib/manager/components/deployment/DeploymentPreviewButton.js
@@ -4,7 +4,7 @@ import Icon from '@conveyal/woonerf/components/icon'
 import React, {Component} from 'react'
 import {Button, Tooltip, OverlayTrigger} from 'react-bootstrap'
 
-import type {Deployment, Project} from '../../types'
+import type {Deployment, Project} from '../../../types'
 
 type Props = {
   deployment: Deployment,
