+++ lib/manager/components/deployment/DeploymentSettings.js
@@ -9,14 +9,14 @@ import {shallowEqual} from 'react-pure-render'
 import {withRouter} from 'react-router'
 import {LinkContainer} from 'react-router-bootstrap'
 
-import FormInput from '../../common/components/FormInput'
-import {getComponentMessages} from '../../common/util/config'
-import CollapsiblePanel from './CollapsiblePanel'
-import {parseBounds} from '../util'
-import {FIELDS, SERVER_FIELDS, UPDATER_FIELDS} from '../util/deployment'
+import FormInput from '../../../common/components/FormInput'
+import {getComponentMessages} from '../../../common/util/config'
+import CollapsiblePanel from '../CollapsiblePanel'
+import {parseBounds} from '../../util'
+import {FIELDS, SERVER_FIELDS, UPDATER_FIELDS} from '../../util/deployment'
 
-import {updateProject} from '../actions/projects'
-import type {Project} from '../../types'
+import {updateProject} from '../../actions/projects'
+import type {Project} from '../../../types'
 
 type Props = {
   editDisabled: boolean,
