+++ lib/manager/components/deployment/DeploymentVersionsTable.js
@@ -4,13 +4,13 @@ import React, {Component} from 'react'
 import {Table, Button, Badge, Glyphicon, ButtonToolbar} from 'react-bootstrap'
 import { Link } from 'react-router'
 
-import * as deploymentActions from '../actions/deployments'
-import {getComponentMessages} from '../../common/util/config'
-import {formatTimestamp, fromNow} from '../../common/util/date-time'
-import {versionHasExpired, versionHasNotBegun} from '../util/version'
+import * as deploymentActions from '../../actions/deployments'
+import {getComponentMessages} from '../../../common/util/config'
+import {formatTimestamp, fromNow} from '../../../common/util/date-time'
+import {versionHasExpired, versionHasNotBegun} from '../../util/version'
 
-import type {Deployment, Feed, Project, SummarizedFeedVersion} from '../../types'
-import type {ManagerUserState} from '../../types/reducers'
+import type {Deployment, Feed, Project, SummarizedFeedVersion} from '../../../types'
+import type {ManagerUserState} from '../../../types/reducers'
 
 type Props = {
   deleteFeedVersion: typeof deploymentActions.deleteFeedVersion,
