+++ lib/manager/components/deployment/DeploymentViewer.js
@@ -28,30 +28,30 @@ import Select from 'react-select'
 import fetch from 'isomorphic-fetch'
 import validator from 'validator'
 
-import * as deploymentActions from '../actions/deployments'
-import Loading from '../../common/components/Loading'
-import EditableTextField from '../../common/components/EditableTextField'
-import Title from '../../common/components/Title'
-import WatchButton from '../../common/containers/WatchButton'
-import {getComponentMessages, getConfigProperty} from '../../common/util/config'
-import {formatTimestamp, fromNow} from '../../common/util/date-time'
-import {isValidJSONC} from '../../common/util/json'
-import {defaultTileLayerProps} from '../../common/util/maps'
-import { versionsSorter } from '../../common/util/util'
-import {getServerDeployedTo} from '../util/deployment'
+import * as deploymentActions from '../../actions/deployments'
+import Loading from '../../../common/components/Loading'
+import EditableTextField from '../../../common/components/EditableTextField'
+import Title from '../../../common/components/Title'
+import WatchButton from '../../../common/containers/WatchButton'
+import {getComponentMessages, getConfigProperty} from '../../../common/util/config'
+import {formatTimestamp, fromNow} from '../../../common/util/date-time'
+import {isValidJSONC} from '../../../common/util/json'
+import {defaultTileLayerProps} from '../../../common/util/maps'
+import { versionsSorter } from '../../../common/util/util'
+import {getServerDeployedTo} from '../../util/deployment'
 import CurrentDeploymentPanel from './CurrentDeploymentPanel'
 import DeploymentConfirmModal from './DeploymentConfirmModal'
 import DeploymentVersionsTable from './DeploymentVersionsTable'
 
-import type {Props as ContainerProps} from '../containers/ActiveDeploymentViewer'
+import type {Props as ContainerProps} from '../../containers/ActiveDeploymentViewer'
 import type {
   CustomFile,
   Deployment,
   ReactSelectOption,
   ServerJob,
   SummarizedFeedVersion
-} from '../../types'
-import type {ManagerUserState} from '../../types/reducers'
+} from '../../../types'
+import type {ManagerUserState} from '../../../types/reducers'
 
 type Props = ContainerProps & {
   addFeedVersion: typeof deploymentActions.addFeedVersion,
