+++ lib/manager/components/deployment/DeploymentsPanel.js
@@ -15,19 +15,19 @@ import {
   Panel
 } from 'react-bootstrap'
 
-import * as deploymentActions from '../actions/deployments'
-import * as feedsActions from '../actions/feeds'
-import * as projectsActions from '../actions/projects'
-import ActiveDeploymentViewer from '../containers/ActiveDeploymentViewer'
-import ConfirmModal from '../../common/components/ConfirmModal'
-import EditableTextField from '../../common/components/EditableTextField'
-import Loading from '../../common/components/Loading'
-import {getComponentMessages} from '../../common/util/config'
-import {formatTimestamp, fromNow} from '../../common/util/date-time'
-import {getServerDeployedTo} from '../util/deployment'
+import * as deploymentActions from '../../actions/deployments'
+import * as feedsActions from '../../actions/feeds'
+import * as projectsActions from '../../actions/projects'
+import ActiveDeploymentViewer from '../../containers/ActiveDeploymentViewer'
+import ConfirmModal from '../../../common/components/ConfirmModal'
+import EditableTextField from '../../../common/components/EditableTextField'
+import Loading from '../../../common/components/Loading'
+import {getComponentMessages} from '../../../common/util/config'
+import {formatTimestamp, fromNow} from '../../../common/util/date-time'
+import {getServerDeployedTo} from '../../util/deployment'
 
-import type {Props as ContainerProps} from '../containers/DeploymentsPanel'
-import type {Deployment, Project} from '../../types'
+import type {Props as ContainerProps} from '../../containers/DeploymentsPanel'
+import type {Deployment, Project} from '../../../types'
 
 type Props = ContainerProps & {
   createDeployment: typeof deploymentActions.createDeployment,
